requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
asgiref==3.8.1
attrs==24.2.0
certifi==2024.7.4
charset-normalizer==3.3.2
discord.py==2.4.0
Django==5.1.2
django-cors-headers==4.5.0
django-dotenv==1.4.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
frozenlist==1.4.1
idna==3.7
multidict==6.0.5
mypy==1.11.1
mypy-extensions==1.0.0
PyJWT==2.9.0
requests==2.32.3
setuptools==73.0.1
sqlparse==0.5.1
typing_extensions==4.12.2
urllib3==2.2.2
wheel==0.44.0
whitenoise==6.7.0
yarl==1.9.4
